Move SchameNameAPI to GetSchemaAPI
