+++ solr/core/src/test/org/apache/solr/handler/admin/api/GetSchemaAPITest.java
@@ -21,7 +21,11 @@ import static org.mockito.Mockito.when;
 
 import java.util.Map;
 import org.apache.solr.SolrTestCaseJ4;
+import org.apache.solr.client.solrj.response.schema.SchemaResponse;
+import org.apache.solr.common.util.NamedList;
 import org.apache.solr.common.util.SimpleOrderedMap;
+import org.apache.solr.handler.SchemaHandler;
+import org.apache.solr.handler.api.V2ApiUtils;
 import org.apache.solr.schema.IndexSchema;
 import org.apache.solr.schema.SchemaField;
 import org.apache.solr.schema.SimilarityFactory;
@@ -53,6 +57,39 @@ public class GetSchemaAPITest extends SolrTestCaseJ4 {
     assertNotNull(response.schema);
     assertEquals(1, response.schema.size());
     assertEquals("flagValue", response.schema.get("flagKey"));
+  }
+
+  @Test
+  public void testLooksUpNameFromLatestCoreSchema() throws Exception {
+    when(mockSchema.getSchemaName()).thenReturn("expectedSchemaName");
+
+    final GetSchemaAPI.SchemaNameResponse response = api.getSchemaName();
+
+    assertEquals("expectedSchemaName", response.name);
+    assertNull(response.error);
+  }
+
+  /**
+   * Test the v2 to v1 response mapping for /schema/name
+   *
+   * <p>{@link SchemaHandler} uses the v2 {@link GetSchemaAPI} (and its response class {@link
+   * GetSchemaAPI.SchemaNameResponse}) internally to serve the v1 version of this functionality.
+   * So it's important to make sure that our response stays compatible with SolrJ - both because
+   * that's important in its own right and because that ensures we haven't accidentally changed the
+   * v1 response format.
+   */
+  @Test
+  public void testResponseCanBeParsedBySolrJ() {
+    final NamedList<Object> squashedResponse = new NamedList<>();
+    final GetSchemaAPI.SchemaNameResponse typedResponse =
+      new GetSchemaAPI.SchemaNameResponse();
+    typedResponse.name = "someName";
+
+    V2ApiUtils.squashIntoNamedList(squashedResponse, typedResponse);
+    final SchemaResponse.SchemaNameResponse solrjResponse = new SchemaResponse.SchemaNameResponse();
+    solrjResponse.setResponse(squashedResponse);
+
+    assertEquals("someName", solrjResponse.getSchemaName());
   }
 
   @Test
